Add rectangle movement calculation

+++ include/model/Rectangle.hpp
@@ -7,14 +7,14 @@
 #include "Entity.hpp"
 
 class Rectangle: public Entity {
-    int rectangle_width;
-    int rectangle_height;
+    float rectangle_width;
+    float rectangle_height;
 
 public:
-    Rectangle(Vector2d position, int rotation, Vector2d velocity, double speed, int rectangle_width, int rectangle_height);
+    Rectangle(Vector2d position, int rotation, Vector2d velocity, double speed, float rectangle_width, float rectangle_height);
 
-    [[nodiscard]] int getWidth() const;
-    [[nodiscard]] int getHeight() const;
+    [[nodiscard]] float getWidth() const;
+    [[nodiscard]] float getHeight() const;
 
 };
